+++ test/function/arithmetic/edivide.test.js
@@ -51,8 +51,8 @@ describe('edivide', function() {
   });
 
   it('should divide all the elements of a matrix by one number', function() {
-    assert.deepEqual(edivide(math.range(2,8,2), 2), [1,2,3]);
-    a  = math.matrix([[1,2],[3,4]]);
+    assert.deepEqual(edivide([2,4,6], 2), [1,2,3]);
+    var a = math.matrix([[1,2],[3,4]]);
     assert.deepEqual(edivide(a, 2), math.matrix([[0.5,1],[1.5,2]]));
     assert.deepEqual(edivide(a.valueOf(), 2), [[0.5,1],[1.5,2]]);
     assert.deepEqual(edivide([], 2), []);