+++ test/function/matrix/ones.test.js
@@ -1,7 +1,8 @@
 // test ones
 var assert = require('assert');
 var math = require('../../../index.js'),
-    ones = math.ones;
+    ones = math.ones,
+    matrix = math.matrix;
 
 describe('ones', function() {
 
@@ -11,27 +12,29 @@ describe('ones', function() {
   });
 
   it('should create a vector with ones', function () {
-    assert.deepEqual(ones(3), [1,1,1]);
+    assert.deepEqual(ones(3), matrix([1,1,1]));
+    assert.deepEqual(ones(matrix([4])), matrix([1,1,1,1]));
     assert.deepEqual(ones([4]), [1,1,1,1]);
-    assert.deepEqual(ones(0), []);
+    assert.deepEqual(ones(0), matrix([]));
   });
 
-  it('should create a 2D matrix with ones from an array', function () {
-    assert.deepEqual(ones(2,3), [[1,1,1],[1,1,1]]);
-    assert.deepEqual(ones(3,2), [[1,1],[1,1],[1,1]]);
+  it('should create a 2D matrix with ones', function () {
+    assert.deepEqual(ones(2,3), matrix([[1,1,1],[1,1,1]]));
+    assert.deepEqual(ones(3,2), matrix([[1,1],[1,1],[1,1]]));
     assert.deepEqual(ones([3,2]), [[1,1],[1,1],[1,1]]);
+    assert.deepEqual(ones(matrix([3,2])), matrix([[1,1],[1,1],[1,1]]));
   });
 
   it('should create a matrix with ones from a matrix', function () {
-    assert.deepEqual(ones(math.matrix([3])), math.matrix([1,1,1]));
-    assert.deepEqual(ones(math.matrix([3,2])), math.matrix([[1,1],[1,1],[1,1]]));
+    assert.deepEqual(ones(matrix([3])), matrix([1,1,1]));
+    assert.deepEqual(ones(matrix([3,2])), matrix([[1,1],[1,1],[1,1]]));
 
     // TODO: do we want to support the following? maybe better not
-    assert.deepEqual(ones(math.matrix([[[3]],[[2]]])), math.matrix([[1,1],[1,1],[1,1]]));
+    assert.deepEqual(ones(matrix([[[3]],[[2]]])), matrix([[1,1],[1,1],[1,1]]));
   });
 
   it('should create a 3D matrix with ones', function () {
-    assert.deepEqual(ones(2,3,4), [
+    var res = [
       [
         [1,1,1,1],
         [1,1,1,1],
@@ -42,23 +45,16 @@ describe('ones', function() {
         [1,1,1,1],
         [1,1,1,1]
       ]
-    ]);
-    assert.deepEqual(ones([2,3,4]), [
-      [
-        [1,1,1,1],
-        [1,1,1,1],
-        [1,1,1,1]
-      ],
-      [
-        [1,1,1,1],
-        [1,1,1,1],
-        [1,1,1,1]
-      ]
-    ]);
+    ];
+    assert.deepEqual(ones(2,3,4), matrix(res));
+    assert.deepEqual(ones(matrix([2,3,4])), matrix(res));
+    assert.deepEqual(ones([2,3,4]), res);
   });
 
+  // TODO: test option math.options.matrix.default
+
   it('should create a matrix with ones with the same size as original matrix', function () {
-    var a = math.matrix([[1, 2, 3], [4, 5, 6]]);
+    var a = matrix([[1, 2, 3], [4, 5, 6]]);
     assert.deepEqual(ones(math.size(a)).size(), a.size());
   });